revised file and folder structure
