+++ src/common/Card/Reminder/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Reminder";
