+++ src/common/Card/Reminder/style.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const StyledReminder = styled.div`
+  background: #ff7a7a;
+  color: white;
+  font-size: 0.7em;
+  font-weight: bold;
+  padding: 0.5em;
+  position: absolute;
+  top: 1em;
+  left: 1em;
+  border-radius: 8px;
+`;
